moved components out of friendsview
